Move RandomClipSampler to references
